pyproject.toml
[project]
name = "pp-agent-toolkit"
version = "0.2.0"
description = "Power Platform Copilot Studio agent toolkit: rename solution exports, visualise agent structure, and validate system instructions against model best practices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "reflex>=0.6.0",
    "rich>=13.9.0",
    "typer>=0.12.0",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
